Remove Proxy overhead with two-phase agent initialization
  The agent previously used java.lang.reflect.Proxy to bridge between
  classloaders when setting up the enforcement callback. This added
  unnecessary overhead in the initialization path.

  The fix ensures single type identity for bootstrap classes at runtime:

  1. agent-bootstrap dependency changed to compileOnly (not packaged)
  2. shadowJar excludes org/jguard/bootstrap/** classes
  3. Bootstrap JAR embedded at jguard/bootstrap.jar and injected first
  4. AgentInitializer loaded via reflection AFTER bootstrap injection

  This solves the classloader chicken-and-egg problem:
  - JGuardAgent.premain() cannot import bootstrap types (they don't exist yet)
  - After injecting bootstrap.jar, AgentInitializer is loaded via reflection
  - AgentInitializer can freely use bootstrap types with direct lambda

  The enforcement callback is now a simple lambda instead of a Proxy,
  eliminating reflection overhead in the hot path.

  Files changed:
  - agent/build.gradle: compileOnly + testImplementation + exclusions
  - JGuardAgent.java: Minimal entry point, no bootstrap imports
  - AgentInitializer.java: New class with real initialization logic
  - CONTRIBUTING.md: Updated Key Classes table and Step 4 reference
